move to Helpers instead of Connect
